% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_query_list}
\alias{create_query_list}
\title{Create query list}
\usage{
create_query_list(...)
}
\arguments{
\item{...}{Name value pairs.}
}
\value{
A named list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
