% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popular.R
\name{get_most_emailed}
\alias{get_most_emailed}
\title{Get most emailed articles}
\usage{
get_most_emailed(period = 30, key = NULL, parse = TRUE)
}
\arguments{
\item{period}{Numeric. Period to get data. Either 1, 7 or 30. Defaults to 30.}

\item{key}{String. NYT key.}

\item{parse}{Logical. TRUE to return parsed JSON, FALSE to return HTTP
response.}
}
\value{
HTTP response or parsed JSON.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
