% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytapi-package.R
\docType{package}
\name{nytapi-package}
\alias{nytapi}
\alias{nytapi-package}
\title{nytapi: Get Data from the New York Times API}
\description{
Get data from the New York Times API.
}
\author{
\strong{Maintainer}: Judith Bourque \email{info@clessn.ca}

}
\keyword{internal}
