% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_articles}
\alias{search_articles}
\title{Look up articles by keywords}
\usage{
search_articles(query = NULL, filter = NULL, key = NULL)
}
\arguments{
\item{query}{Search query}

\item{filter}{Search filters}

\item{key}{API key}
}
\value{
HTTR response
}
\description{
Look up articles by keywords
}
