% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popular.R
\name{search_articles}
\alias{search_articles}
\title{Get most popular articles}
\usage{
search_articles(query = NULL, filter = NULL, ..., key = NULL)
}
\arguments{
\item{query}{Search query}

\item{filter}{Search filters}

\item{...}{Other things to pass onto query}

\item{key}{API key}
}
\value{
HTTR response
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
