% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_articles}
\alias{search_articles}
\title{Look up articles by keywords}
\usage{
search_articles(query = NULL, filter = NULL, ..., key = NULL, parse = TRUE)
}
\arguments{
\item{query}{Search query}

\item{filter}{Search filters}

\item{...}{Other things to pass onto query}

\item{key}{API key}

\item{parse}{Logical. TRUE to return parsed JSON, FALSE to return HTTP
response.}
}
\value{
HTTP response or parsed JSON.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
